+++ .github/workflows/update_protobufs.yml
@@ -37,5 +37,5 @@ jobs:
           git config --global user.email 'bot@noreply.github.com'
           git remote set-url origin https://x-access-token:${{ secrets.GITHUB_TOKEN }}@github.com/${{ github.repository }}
           git add protobufs
-          git add meshtastic
+          git add meshtastic/protobuf
           git commit -m "Update protobuf submodule" && git push || echo "No changes to commit"
