+++ bin/regen-protobufs.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+set -e
+
 #Uncomment to run hack
 #gsed -i 's/import "\//import ".\//g' ./protobufs/meshtastic/*
 #gsed -i 's/package meshtastic;//g' ./protobufs/meshtastic/*
@@ -7,17 +9,45 @@
 # protoc looks for mypy plugin in the python path
 source $(poetry env info --path)/bin/activate
 
-./nanopb-0.4.8/generator-bin/protoc -I=protobufs --python_out ./ --mypy_out ./ ./protobufs/meshtastic/*.proto
-./nanopb-0.4.8/generator-bin/protoc -I=protobufs --python_out ./meshtastic/ --mypy_out ./meshtastic/ ./protobufs/nanopb.proto
+# Put our temp files in the poetry build directory
+TMPDIR=./build/meshtastic/protofixup
+echo "Fixing up protobuf paths in ${TMPDIR} temp directory"
 
-# workaround for import bug in protoc https://github.com/protocolbuffers/protobuf/issues/1491#issuecomment-690618628
 
+# Ensure a clean build
+rm -r "${TMPDIR}"
+
+INDIR=${TMPDIR}/in/meshtastic/protobuf
+OUTDIR=${TMPDIR}/out
+PYIDIR=${TMPDIR}/out
+mkdir -p "${OUTDIR}" "${INDIR}" "${PYIDIR}"
+cp ./protobufs/meshtastic/*.proto "${INDIR}"
+
+# OS-X sed is apparently a little different and expects an arg for -i
 if [[ $OSTYPE == 'darwin'* ]]; then
-	sed -i '' -E 's/^(import.*_pb2)/from . \1/' meshtastic/*.py
-	# automate the current workaround (may be related to Meshtastic-protobufs issue #27 https://github.com/meshtastic/protobufs/issues/27)
-	sed -i '' -E "s/^None = 0/globals()['None'] = 0/" meshtastic/mesh_pb2.py
+	SEDCMD="sed -i '' -E"
 else
-	sed -i -e 's/^import.*_pb2/from . \0/' meshtastic/*.py
-	# automate the current workaround (may be related to Meshtastic-protobufs issue #27 https://github.com/meshtastic/protobufs/issues/27)
-	sed -i -e "s/^None = 0/globals()['None'] = 0/" meshtastic/mesh_pb2.py
+	SEDCMD="sed -i -E"
 fi
+
+
+# change the package names to meshtastic.protobuf
+$SEDCMD 's/^package meshtastic;/package meshtastic.protobuf;/' "${INDIR}/"*.proto
+# fix the imports to match
+$SEDCMD 's/^import "meshtastic\//import "meshtastic\/protobuf\//' "${INDIR}/"*.proto
+
+# Generate the python files
+./nanopb-0.4.8/generator-bin/protoc -I=$TMPDIR/in --python_out "${OUTDIR}" "--mypy_out=${PYIDIR}" $INDIR/*.proto
+
+# Change "from meshtastic.protobuf import" to "from . import"
+$SEDCMD 's/^from meshtastic.protobuf import/from . import/' "${OUTDIR}"/meshtastic/protobuf/*pb2*.py[i]
+
+# Create a __init__.py in the out directory
+touch "${OUTDIR}/meshtastic/protobuf/__init__.py"
+
+# Copy to the source controlled tree
+mkdir -p meshtastic/protobuf
+rm -rf meshtastic/protobuf/*pb2*.py
+cp "${OUTDIR}/meshtastic/protobuf"/* meshtastic/protobuf
+
+exit 0
