+++ meshtastic/__init__.py
@@ -81,7 +81,10 @@ from google.protobuf.json_format import MessageToJson
 from pubsub import pub # type: ignore[import-untyped]
 from tabulate import tabulate
 
-from meshtastic import (
+from meshtastic.node import Node
+from meshtastic.util import DeferredExecution, Timeout, catchAndIgnore, fixme, stripnl
+
+from .protobuf import (
     admin_pb2,
     apponly_pb2,
     channel_pb2,
@@ -93,10 +96,10 @@ from meshtastic import (
     remote_hardware_pb2,
     storeforward_pb2,
     telemetry_pb2,
+)
+from . import (
     util,
 )
-from meshtastic.node import Node
-from meshtastic.util import DeferredExecution, Timeout, catchAndIgnore, fixme, stripnl
 
 # Note: To follow PEP224, comments should be after the module variable.
