+++ meshtastic/remote_hardware.py
@@ -4,7 +4,7 @@ import logging
 
 from pubsub import pub # type: ignore[import-untyped]
 
-from meshtastic import portnums_pb2, remote_hardware_pb2
+from meshtastic.protobuf import portnums_pb2, remote_hardware_pb2
 from meshtastic.util import our_exit
